requirements.txt
Django==3.2
djangorestframework==3.12.4
djongo==1.3.6
pymongo==3.11.4
python-dotenv==0.15.0
pytz==2021.1
openai
